requirements.txt
feedparser==6.0.10
discord.py==2.2.3
openai==0.27.4
python-dotenv==1.0.0
